Improve detection of Romaji in TMDB titles
This is going to be an uphill battle every time since none of the
data actually lines up. It's unfortunate, but this is probably the
next best step I can take.

There are a few more cases that aren't currently handled:

* No alternative title, but also no translated title
* No Japanese alternative title, but other ones such as English have it

The romaji conversion routine was supposed to help with this but it's
currently completely unused because it needs more research in order
to be used.

Right now this commit improves the detection for the ASCII case
where the title has macrons.