Add support for notifications
This adds support for two features, bookmarking an entry and then
getting notified when that entry gets new files.

I've explicitly (for now) decided against notifying people when a
kitsunekko or tmdb scrape happens to upload files though I might change
my mind in the future.

In the future this notification feature will hopefully be used to notify
people of other moderator related events such as reports or when an
admin wants to notify a user (or multiple users) about something.

Resolves #28